+++ data-raw/generate_themes.R
@@ -436,29 +436,21 @@ dd <- convert_to_rstudio_theme(
 )
 
 # Positron -------------------------------------------------------------------
-url <- file.path(
-  "https://raw.githubusercontent.com/posit-dev/positron/refs/heads/main",
-  "/extensions/theme-2026/themes/2026-dark.json"
-)
 # Generate rstheme
 dd <- convert_to_rstudio_theme(
-  url,
-  "inst/rsthemes/Positron_2026_Dark.rstheme",
-  name = "Positron 2026 Dark",
+  "data-raw/vscode_themes/positron_dark.json",
+  "inst/rsthemes/Positron_Dark.rstheme",
+  name = "Positron Dark",
   force = TRUE,
   apply = TRUE,
   output_style = "compact"
 )
 
-url <- file.path(
-  "https://raw.githubusercontent.com/posit-dev/positron/refs/heads/main",
-  "/extensions/theme-2026/themes/2026-light.json"
-)
 # Generate rstheme
 dd <- convert_to_rstudio_theme(
-  url,
-  "inst/rsthemes/Positron_2026_Light.rstheme",
-  name = "Positron 2026 Light",
+  "data-raw/vscode_themes/positron_light.json",
+  "inst/rsthemes/Positron_Light.rstheme",
+  name = "Positron Light",
   force = TRUE,
   apply = TRUE,
   output_style = "compact"